backend/requirements.txt
Django==4.0.3
django-cors-middleware==1.5.0
django-extensions==3.1.5
djangorestframework==3.13.1
PyJWT==2.3.0
psycopg2==2.9.3
django-environ==0.8.1
